+++ application/dash_application/views/single_lane_rna.py
@@ -11,10 +11,12 @@ from ..plot_builder import fill_in_colour_col, fill_in_shape_col, \
 from ..table_builder import table_tabs, cutoff_table_data
 from ..utility import df_manipulation as util
 from ..utility import sidebar_utils
+from ..utility import log_utils
 from gsiqcetl.column import RnaSeqQcColumn as RnaColumn
 import pinery
 import logging
 import json
+import datetime
 
 logger = logging.getLogger(__name__)
 
@@ -169,8 +171,18 @@ ALL_TISSUE_MATERIALS = RNA_DF[
     PINERY_COL.TissuePreparation].sort_values().unique()
 ALL_LIBRARY_DESIGNS = RNA_DF[
     PINERY_COL.LibrarySourceTemplateType].sort_values().unique()
-ALL_RUNS = RNA_DF[RNA_COL.Run].sort_values().unique()[::-1]  # reverse the list
+ALL_RUNS = RNA_DF[PINERY_COL.SequencerRunName].sort_values().unique()[::-1]  # reverse the list
 ALL_SAMPLES = RNA_DF[PINERY_COL.SampleName].sort_values().unique()
+
+# N.B. The keys in this object must match the argument names for
+# the `update_pressed` function in the views.
+collapsing_functions = {
+    "projects": lambda selected: log_utils.collapse_if_all_selected(selected, ALL_PROJECTS, "all_projects"),
+    "runs": lambda selected: log_utils.collapse_if_all_selected(selected, ALL_RUNS, "all_runs"),
+    "kits": lambda selected: log_utils.collapse_if_all_selected(selected, ALL_KITS, "all_kits"),
+    "instruments": lambda selected: log_utils.collapse_if_all_selected(selected, ILLUMINA_INSTRUMENT_MODELS, "all_instruments"),
+    "library_designs": lambda selected: log_utils.collapse_if_all_selected(selected, ALL_LIBRARY_DESIGNS, "all_library_designs"),
+}
 
 shape_or_colour_values = {
     PINERY_COL.StudyTitle: ALL_PROJECTS,
@@ -508,9 +520,7 @@ def init_callbacks(dash_app):
                        start_date,
                        end_date,
                        search_query):
-        params = locals()
-        del params['click']
-        logger.info(json.dumps(params))
+        log_utils.log_filters(locals(), collapsing_functions, logger)
         
         if not runs and not instruments and not projects and not kits and not library_designs:
             df = EMPTY_RNA
@@ -518,7 +528,7 @@ def init_callbacks(dash_app):
             df = RNA_DF
 
         if runs:
-            df = df[df[RNA_COL.Run].isin(runs)]
+            df = df[df[PINERY_COL.SequencerRunName].isin(runs)]
         if instruments:
             df = df[df[INSTRUMENT_COLS.ModelName].isin(instruments)]
         if projects:
@@ -528,7 +538,7 @@ def init_callbacks(dash_app):
         if library_designs:
             df = df[df[PINERY_COL.LibrarySourceTemplateType].isin(
                 library_designs)]
-        df = df[df[RNA_COL.Run].isin(sidebar_utils.runs_in_range(start_date, end_date))]
+        df = df[df[PINERY_COL.SequencerRunName].isin(sidebar_utils.runs_in_range(start_date, end_date))]
         sort_by = [first_sort, second_sort]
         df = df.sort_values(by=sort_by)
         df = fill_in_shape_col(df, shape_by, shape_or_colour_values)