+++ CHANGELOG.md
@@ -10,8 +10,14 @@ and as of version 1.0.0, follows semantic versioning.
   * Bcl2Fastq Index QC graph
   * Set `DASHI_LOG_TO_CONSOLE=True` in `.env` file to log to console
   * Set `USE_BLEEDING_EDGE_ETL=1` in `.env` file to use gsi-qc-etl@master (development only)
+### Changed
+  * Filter logs now remove `end_date` if it is the current date
+  * Filter logs now report `end_date` as a date rather than datetime
+  * Filter logs now report `["all_runs"]` when all runs have been selected; etc for other dropdowns
 ### Fixed
   * `nav_handler` and `content_handler` no longer throw exception on empty path
+  * `Add All` button for Library Designs on WGS report now works
+  * Remove blank Run from TS and RNA dropdowns
 
 ## [200114-1651] - 2020-01-14
 ### Added